Add option to speed-up remote review with checkbox.
Default is false, which is the old behavior using E to go to the next
segment. If enabled (true), after 2 times pressing W it automatically
jumps to the next unreviewed segment, sets the actual segment's review
status to 100.00 on the client side and does not refresh the widget,
thus speeding up reviewing of the long tail of very short segments.