Annotations: improve query performance for last used and usage fields
This commit changes the extra last used and usage count fields of an annotation
query to explicit SQL. This improves the query performance considerably. That is
because the Django syntax used before referred to the joined tables which are
much larger than the single tables actually only required which are used now.