Annotation navigator: support 'annotates' query in annotation node
If one selects an annotation in the navigator, one is presented an
annotation node. This node provides options what one can do next, based
on the annotation chosen. With this commit a new option is introduced:
'Annotates'. When selected a new annotation list node will be displayed
that shows only annotations that are annotated with the last annotation.

Vice versa, the 'Annotated with' option, will only show annotations that
are annotating the currently selected one.