Annotations: migrate skeleton lists to annotations
The first migration included in this commit will go though all skeleton lists to
create an annotation for each one (if none with the same name exists already).
The neuron of every skeleton referenced by a list will then be linked to this
annotation.

The second migration will remove the skeleton table alltogether. Additionally,
this commit removes the save and load list functionality from the staging area.