Annotation navigator: implement skeleton source interface in neuron node
If a neuron node is loaded (e.g. when the active node is selected in the home
node), it can now highlight the skeleton row if the very same skeleton is
selected in another source. It can also provide information about the currently
selected skeleton and provide models for it.