Join skeletons: add better one-node skeleton handling
If the to-skeleton has only one node, it is checked whether is has any
annotations. If not, the node is joined right away. Otherwise, the dialog is
shown. If the to-skeleton has more than one node, the dialog is shown in all
cases.

For the sake of simplicity, the annotations of from and to neurons are
potentially retrieved more than once in the case of single to-node: If it turns
out that dialog should be shown for a single to-node (because it is annotated),
the split/merge dialog will retrieve the annotations again. Additionally, the
single to-node case retrieves also the from-node annotations.