Annotations: get neuron count and queried neurons in separate queries
Since it is very likely that there are many neurons, it is more
efficient to do two queries: 1. Get total number of neurons 2. Get
limited set. The alternative would be to get all neurons for counting
and limiting on the Python side. This, however, is too expensive when
there are many neurons.