Annotations: prevent de-annotation without permissions on link
Up to to now the de-annotation was forbidden when a user has no permissions on
the entity itself (e.g. a neuron). This, however, allows everyone with
permissions on the neuron (e.g. if it is unlocked) to remove every annotation.
This commit changes this behavior to only allow the removal of an annotation if
a user has permissions on it's specific link.