Annotation search: improve performance of selected skeleton retrieval
A new field has been added that maps entity IDs to their selection state. This
is kept up to date on checkbox click events. When selected entities need to be
retrieved, this map can now be used instead of querying the DOM for the check
state of every checkbox. This improves speed especially for large result lists
considerably.