Neuron navigator: replace sub-neuron-list-node with mixin
A node isn't supposed to have a sub-neuron as a field, because it
introduces problems when using deepCopy() for cloning the navigator
window. Therefore, the neuron list nodes used in the home and the
annotation nodes (introduced in 895d5b8) are replaced by a more
re-usable mixin datastructure. Now all nodes that want to display a
neuron list, should include this mixin. The mixin supports highlighting
and retrieval of selected skeletons.