Navigator: make active neuron node recreate filter on update
The active neuron node actively modifies the current filter set if it is
updated. However, because the currently active neuron is evaluated during
content creation, the filter set has been created without the information of a
newly selected neuron. Therefore, filter needs to be updated with the new neuron
ID. To be on the safe side, this commit makes sure the complete filter is
updated.