Navigator: do co-annotation filtering client side
When options for co-annotations are displayed, the list of options doesn't
include co-annotations that have been selected previously. This is to have
a more straightforward UI, because it would make no sense to re-select
already selected annotations.

This commit moves the filtering needed for this from the back-end to the
front-end. The back-end should provide a clean API and wrt. co-annotations
it wasn't clean if co-annotations were removed from a result set. These are
co-annotations to each other as well, after all. Now the front-end does the
filtering where it wants it: when displaying a list of options for
co-annotations.