Annotations: clean-up back-end method for adding annotations
Respect the fact that not only neurons can be annotated. This also fixes a
problem that an annotation link was added multiple times when meta-annotations
were added. This resulted from re-using existing annotation links only when
annotations were linked to a neuron. This constraint has been removed and the
existence test now uses only user, relation and the IDs of the annotation and
the annotated entity (and not their classes).