Navigator: add button and functionality to neuron node to delete neuron
Deletion of a neuron is only possible if a user is super-user or (s)he has
permissions on all treenodes on all skeletons modeling the neuron as well as on
all annotations linked to it. If this is the case, a neuron, it' skeletons and
treenodes, connectors, treenoded-class-instance-links and annotations are
removed.