Add option "highlight_tags" to Tag project view
There is now a config option "highlight_tags" for the Project tag table which
expects a list of tags (just like filter_tags, row_tags and col_tags do). In
the data view, all projects tagged with any of the tags provided will be
highlighted in bold (#dataview .highlight in screen.css).