Annotation search: make annotation input field ID a class attribute
This field can be duplicated with the help of the "+" button next to it. When
doing this, the input elements' ID isn't updated. Multiple occurrences of the
same ID is illegal and since it will soon be required to address all of these
input fields in one go, the ID has been replaced by a class attribute with the
same value.