Annotations: return used annotations when adding new links
When new annotations are added, the back-end now returns a list of all
annotation objects that have been used for the requested links. In particular
the names and IDs of those annotations are returned. The front-end builds its
status message now from the annotations returned instead of the ones from the
request.