Remove redundant and unused static admin files
In Django, static files are collected from every application with the help of
the 'collectstatic' management command. They are linked or copied in the folder
specified by the STATIC_ROOT variable. Django's admin application brings it's
own static files in it's own folders so there is no need to have a duplicate
version of it.