Remove unneeded configuration files
This commits remove *_production and *_local configuration files. If a
user needs an additional configuration it is easy enough to do this
based on the existing settings.py file. Also, if a CATMAID server should
be used as production system, one should probably spend a minute or two
on the settings. Therefore, it might be better to not motivate a safe
production setup by creating *_production files. This change relates to
issue #581.