Remove some namespace imports from catmaid.control package
This commit removes the namespace imports of the following modules from
catmaid.control.__init__.py: classification, notifications, roi and
clustering. This is part of an effort to make the URL configuration and
the control package more readable and maintainable. The URL configuration
has been adjusted as well: The corresponding prefixes have been
extended.

I double checked, that all catmaid.control imports removed, where not in
use in other places.

This commit relates to issue #570.