Remove authentication namespace import from catmaid.control package
Besides removing the authentication namespace import from the
catmaid.control package, this commit also adds the authentication prefix
to the URL configuration where required. This is part of an effort to
make the URL configuration and the control package more readable and
maintainable.

I double checked, that all catmaid.control imports removed, where not in
use in other places.

This commit relates to issue #570.