Move treenode access URLs into own statement
This is part of an effort to make the URL configuration and the control
package more readable and maintainable. I double checked, that all
catmaid.control namespace imports removed, where not in use in other
places. Somehow the join_skeletons_interpolated method is also part of the
treenodes module. I prepended it's reference in view with the explicit
module name.

This commit relates to issue #570.