Remove obsolete method in treenode module
The 'update_treenode_table' method has been moved to the treenodetable
module in commit d2c42151. It wasn't removed from the treenode module
back than and this commit fixes this to avoid confusion. I tested the
treenode table widget in the front: it still works. Everything else
would have been surprising, because the URLs controller links the call
to the method in the treenodetable module. It isn't used anywhere else.

This is related to issue #586.