Move skeleton access URLs into own statement
This is part of an effort to make the URL configuration and the control
package more readable and maintainable. I double checked, that all
catmaid.control namespace imports removed, where not in use in other
places.

There are certainly several valid ways to organize the API (not only for
skeletons). For now, I thought, it is best to have bigger sets of
functions from one module in an own statement. If it turns out this is
to confusing, there shouldn't be any problem to group the API more
semantically (e.g. everything that relates to skeletons is in one
statement).

This commit relates to issue #570.