Move skeleton export URLs into own statement
This is part of an effort to make the URL configuration and the control
package more readable and maintainable. I double checked, that all
catmaid.control namespace imports removed, where not in use in other
places. This code is probably removed soon. However, this change was
required to clean the catmaid.control package.

This commit relates to issue #570.