Fix bug that prevented selection of a skeleton/nearest node
A callback is used to select a skeleton or node nearest to a position.
This callback is executed after all layers are redrawn. This, however,
only happens if the callback for each layer-rendering is executed. This
did not happen for a tile layer, if it was invisible. This commit fixes
this by not executing the redraw method for an invisible layer in the
first place.

Eventually, to prevent those bugs, layers shouldn't care whether they
are visible or not. This should be handled by the caller, i.e. the
stack.