Move general URL configuration statement to end of file
The general URL configuration only contains one- and two-liners for
which an own statement wouldn't improve readability. To have a better
structure of the URL configuration file in general, this statement is
moved to the end of the file. By doing this, the module specific
configuration statements come first in the file, followed by the general
statement.

This relates to issue #570.