Graph widget: automatically re-layout when resizing
This commit adds a resize handler to the graph widget. In it a timeout
of 400ms is set to re-layout the graph. This timeout is cleared and
renewed if there is a new resize event within the current timeout.

To make this work robustly, I had to adjust the style of the graph
container: It's overflow setting is now set to hidden. This is required,
because cytoscape.js' redraw can be delayed  (e.g. due to animation).
When the window's size is reduced, it can happen that the cytoscape
canvas is bigger than the container. The default 'auto' setting
then introduces scrollbars, triggering another resize. This somehow
confuses cytoscape.js and causes the graph to disappear.

This relates to issue #593.