Tracing tool's project statistics: manually draw pie chart legend
The project statistics widget is part of CATMAID's tracing tools and can
be opened with a click on the small badge icon. It uses pie charts,
accompanied by a legend. This legend can easily grow to large and
Raphael (which is used to draw chart and legend) has no way of dealing
with this properly. This commit therefore changes the drawing routine to
create the legend manually: Name columns are created if the list grows
too long in Y. Of course, also the X can be maxed out and this is the
reason the SVG's width is now set to 100%.

This fixes issue #597.