Make user profile handling more robust
This commit adds a new user profile function and prototype. When a user
profile object is created it checks if all expected fields are
available and throws an exception if not. The single place where a
user profile object is created can handle the exception and displays an
error dialog in such a case. Not being able to initialize the user
profile is a severe problem after all.

The reason for creating a user profile constructor and prototype is that
there has to be one place where information can be stored whether a
particular user profile option can be modified and made persistent by
a user. This is with respect to issue #590.