Support saving of specific user profile settings from front-end
This commit adds front-end and back-end code to allow regular users the
modifications of certain user profile settings. Currently, this is only
the display of reference lines in a stack. The back-end won't allow any
other setting to be modified through the interface added. To allow more
settings to be modified, they have to be added to the back-end code and
they have to be marked as modifiable in the front-end.

Anonymous users are never allowed to change the user profile of the
anonymous user. However, the back-also doesn't complain if the anonymous
user tries. This makes callback implementation in the front-end simpler.

This fixes issue #590.