Connectivity widget: make current neurons selectable
This commit adds checkboxes to the list of the neurons currently looked
at. If the connectivity widget is used as a skeleton source it will now
return only skeleton models for the current neurons that have their
corresponding checkboxes checked (like in the tables).

Additionally styles are added so that longer neuron names wrap properly.

This fixes issue #584.