Connectivity table: support "side by side" and "successive" table layouts
If the toggle checkbox is unchecked, the tables are shown one after the other,
otherwise the display is as before. To do this, the "container table" had to be
removed. Both tables are now in a separate div container which is floating to
the left. The toggle basically sets the width of these container divs to either
100% or 50%.

This relates to issue #600.