Connectivity widget: support Firefox
For creating the table cells' content the innerText property was used. This
property, however, is not supported by Firefox [1], because writing to innerText
is apparently not W3C compliant. Therefore, I replaces this property with
appending a text node explicitly. This fixes the issue and makes the
connectivity widget work in Firefox, too.

[1] http://www.quirksmode.org/dom/w3c_html.html#t04