Graph widget: don't re-layout automatically
An automatic re-layout according to the currently selected layout is
usually not wanted. One can easily destroy a carefully arranged graph
layout otherwise. This commits makes therefore sure the graph is is
merely redrawn and not changed in layout or zoom.

Technically, however, a re-layout is still needed to update the display
when the widget size is reduced. A redraw somehow isn't enough.
Therefore, a re-layout is scheduled after the widget is resized. It is
parameterized to not zoom and not modify the node positions.