Connectivity widget: add individual synapse count for selected neurons
The synapse count column is currently an aggregation over all synapses
of all selected neurons. This commit changes this by splitting this
column for individual neurons, if there is more than one selected. In
this case, the 'syn count' column will have sub columns for the sum and
the respective synapse count for the individual neurons.

This relates to issue #601.