Connectivity widget: add index number to neuron list
When there is more than one neuron selected, the synapse count columns
are separated to show the counts of the individual neurons and the sum.
The sub-columns are numbered and refer to the neuron list. To make this
clearer, this commit adds this number to every entry in the neuron
list/table is numbered as well.