Connectivity widget: rename threshold columns and make them slimmer
Due to their title, the threshold selecting columns in the neuron list
were relatively wide. This commit makes the title a bit cleaner by
renaming them from "Threshold Up" and "Threshold Down" to "Upstream
Threshold" and "Downstream Threshold". At the same time, the width of
the column is restricted so that a line break is inserted.