Connectivity widget: show hidden partner stats in table footer
The connectivity allows to add various filters. One can, for instance,
add thresholds or one can hide partners with only one node. To make the
display consistent, this commit introduces a footer row to the
connectivity tables.  This footer displays the aggregated synapse
count(s), review percentage and node count of all hidden nodes. It's
naming and coloring makes it clear that it isn't referring to an actual
neuron.