Connectivity widget: append new skeletons to end of list
This commit introduces a new field to the connectivity widget: a list
containing only the skeleton IDs of the neurons currently selected. This
is used to maintain the order of the neuron list. New skeletons are
always appended to the end of the list.

Of course, having a separate field introduces complexity and one needs
to make sure both the list and the actual skeleton data structure is
consistent. To me, however, it appears to be the simpler solution,
because there are only a few changes required. The alternative would be
to have a single list of skeleton data structures. When doing this, one
couldn't use the object notation anymore and one would be required to
refactor many parts of the widget.

This fixes #614.