Review system: move single user review data to new table
This commit adds a migration to copy all reviews that are stored in the
treenode table to the new review table. Each review gets a new entry
there. This data migration might take a while to run, if there are a lot
of reviewed treenodes. Also, it can be undone, if wanted.

This is part of issue #580.