Fix broken display of custom admin views
The parameter order changed in the function we use for registering
custom views with admin plus. Since we updated django-adminplus, which
we use for the custom views, along the Django 1.6 update, we were
affected by this change. This commit fixes the registering of our custom
views to conform to the new signature.