Let isreciprocal field of Relation model default to false
Since Django 1.6 a BooleanField defaults to None instead of False, if a
default value isn't specified. Therefore, this commit adds a default
value to this field to maintain current functionality and to be more
explicit. The isreciprocal field of the Relation model was the only
BooleanField without a default value. This is fixed now.