Tracing statistics: improve performance of review editor queries
Like with the node creation query, it is much faster if one gets the user
information separately. For the 6677982 tree node test set, the review stats
timing went down from 2.4s to 1.6s and from 950ms to 890ms.

To improve performance further, it might be sensible to omit the user name
query on the back-end and use the user information stored at the front-end to
replace the user IDs.

This commit is part of the work for issue #625.