Tracing statistics: make user history date range adjustable
This commit adds two new text boxes two specify a start and an end day
to show the user history for. It can be adjusted with the help of a new
"refresh" button next to it. The start defaults to ten days ago.

This is part of the work for issue #625