Tracing statistics: remove review statistics
It is now possible that more than one user can review a tree node. The review
statistics chart would need to reflect this in some way. Because it isn't clear,
yet, how to do this and whether such a statistic is needed in the first place,
this commit removes it.

This is part of the changes discussed in issue #580.