Classification search: display 'unfolded' ontology features
Due to relations like "is_a" the ontology display used so far (the one of the
ontology widget) is compressed. For every "is_a" relation there is a separate
path. This commit changes the current display to show all possible paths. Right
now only as a list, bit this will probably change in the future to a tree.