Review system: generalize review retrieving method from treenode table
This method returns a defaultdict with treenodes as keys and a list of
reviewer IDs as values. The result set can be reduced by passing a list
of treenodes or a list of skeletons. The as values listed users can be
modified by passing a mapping function (to list e.g. user names instead
of IDs). The treenode table now makes use of this function.

This is related to the changes discussed in issue #580.