Data views: let the tag table data view use tag, stack and project indices
This commit improves the performance of this tag data view considerably.
It now doesn't do a single request for neither every stack collection of
all project nor their tags. All this information can be pulled from the
indices. Especially for setups with many stacks, projects and tags this
will improve performance a lot.