Review widget: add basic support for multi-user review
This commit updates frontend and backend code of the review widget to support
multiple users reviewing treenodes. Each treenode has now a list of reviewer IDs
attached which is empty if there are no reviews. The review table itself shows
always the current user in the first column, because this is usually what one is
interested in. The last column contains a union of the reviews of all users
having contributed to the neuron or a specific segment, respectively. If no
reviews have been done, yet, one the selected neuron, only one column (the one
for the current user) is shown and the union column is omitted.

This commit is part of the work discussed in issue #580.