Classification editor: add support for ROIs on root node
This commit adds frontend and backend code to allow users to link
regions of interest (ROIs) to a root node of a classification graph.
Before, this was only possible for children of the root node. But since
the root is a class_instance just as the children are and ROIs are
linked to these, there was not much to change to make this work.

Having ROIs on the root node can be useful for instance to create
overview images for a particular classification graph. Thanks to
Sebastian Dunst for this suggestion.