Review widget: add follower checkboxes to select next segment
This commit introduces a new feature to the review widget: Every reviewer column
gets now a checkbox added in front of the user name. If checked, one "follows"
this user. This only has an effect for selecting the next segment (e.g. by
pressing 'E') during/after the review of one segment. The behavior up to now was
that the next segment is selected, that hasn't been completely reviewed by
oneself. This is now extended to have the next segment selected that has not
been completely reviewed by any followed reviewer. By default this is only
oneself, so default net effect is the same as before. It can, however be
extended to express for instance the trust in the reviews of another person by
skipping segments reviewed by her/him and continuing with a segment that has not
been completely reviewed by any followed user.

This change is part of the review system update discussed in issue #580.