Fix redraw callback handling in ROI and box selection layers
These two layers didn't accept a callback. This broke the callback
handling of the complete stack, because the semaphore wouldn't be
counted down for these layers. This in turn didn't synchronize the
movement between all open stacks if one of these layers were active
(e.g. when cropping or a ROI was displayed).