Connectivity widget: support multi-user reviews
This commit introduces a review mode drop-down to the connectivity widget. It
allows to select what review numbers should be displayed in the incoming and
outgoing tables. If set to "union", the review percentage is displayed,
independent of reviewer. Alternatively, one can select one of the reviewing
users. If selected, only the percentage of the reviewing user is displayed.

This also introduces some changes in the back-end and the response data
structure had to be adjusted slightly. There are probably some performance
improvements possible.

This commit is part of the changes discussed in issue #580.