Fix Django 1.6 update bug in tag based classification graph linker
Before Django 1.6 the session store used pickle to serialize objects.
This has been change to JSON, due to security issues with pickle [1].
The tag based classification graph linker, however, wasn't updated to
make the variables it stores in a session JSON serializable (e.g. sets).
As a result, this tool stopped working. This commit fixes this.

[1] https://docs.djangoproject.com/en/1.6/topics/http/sessions/#session-serialization