3D viewer: attach listeners after initialization
The CLOSE and RESIZE listeners are now attached to the 3D viewer window
after the window and the 3d canvas have been initialized. This prevents
a call to the 3D viewers init() method by the resize handler which
doesn't know and pass the 3D viewer's canvas ID to initialization
method. Moreover, it caused the 3D viewer to be initialized twice. So
maybe this also fixes the 3D viewer destruction issues. Due to the
first initialization before, the canvas element was added to a more or
less random position (the canvas' ID wasn't passed!) so it ended up
in the first project menu entry (because it was the first match).