3D viewer: cancel resize if not initialized
This return replaces the call to init() in case of a missing
initialization. The call to init() wasn't correct, because the
canvas ID was missing. Since this is not known in the resizeView()
method before init() was called with the proper ID, a call to
init() isn't possible in resizeView(). Therefore, this commit
makes the method abort if not initialized---this case shouldn't
happen in the first place.