3D viewer: support own and union review coloring
This commit makes the compact-json API return a mapping of all reviewed
treenodes for the selected skeleton to their reviewer IDs. This is used
by the 3D viewer to allow two new coloring modes "All Reviwed" and "Own
Reviewed". The current user's color is used as a highlighting color in
both cases. A dark grey is used as "unreviewed" color. This replaces the
previous "By Reviewer" coloring mode to support the new multi-user
review system.

The compact-json API still returns the original reviewer ID. It will be
removed in a separate commit.

This is part of the work discussed in issue #580.