Multi user review: remove review field from compact-json query
The new multi user review system moves the review information in its
own table and the review_id column will eventually be removed from the
treenode table. This commit changes the treenode table query of the
compact-json API to not query and return the reviewer_id in a treenode
entry anymore. The review information will be added as a separate object
in the response which contains all reviewer_ids for each treenode of the
current skeleton.

This implies a change of all places where the compact-json is used. This
is the 3D Viewer, the Neuron Navigator and a remote access script. The
former has been changed in the commit before this one and is now only
changed to reflect the index change in the returned fields (the reviewer
field is now removed). The Navigator had a column showing the review
percentage for a neuron node. This column can't be expressed that easily
anymore (a union/own percentage would be possible, though). Is is
therefore removed for now. And the latter just needed to have the indices
and the review retrieval adjusted as well.

Issue #580 is where this and other related changes are discussed.