Graph widget: support union and own review coloring modes
To support multi user reviews, the graph widget has now two new review coloring
modes: by union and by own reviews. These replace the review coloring from
before. The back-end had to be changed for this to work. It now gets the review
information from the new review table.

This commit is part of the changes discussed in issue #580.