Treenode table: deactivate sorting of reviewer column
Since it isn't clear, yet, how sorting of the reviewers column should work in
the treenode table when there are multiple reviewers per node, sorting is
disabled for now. All reviewers of a node are listed in this column, though.
This change has been made in the course of removing all references to the
reviewer_id field of the treenode table to be to remove it soon.

This change is part of the discussion in issue #580.