Tree utils: prevent unwanted change in length of list of extra fields
The props list contains the names of user definable fields that are queried in
addition to some standard fields. The length of this list is used to map them to
the results returned by the database. However, before this list got overridden
in every iteration and since its length was changed to include the standard
fields, the length was wrong in the next iteration. This commit fixes this by
using a separate variable for the resulting mapping. This leaves the original
props list untouched.