Remove hack to create user profile for anonymous user
With commit c384f9b9 we introduced a hack that monkey patches django-guardian to
create a user profile when it creates the anonymous user. Along with the upgrade
of CATMAID to Django 1.6, django-guardian had to be upgraded to 1.2 as well for
compatibility reasons. This apparently makes the user profile hack obsolete and
lets it in fact create problems. When django-guardian creates the anonymous
user, our monkey patch code is called, but the user profile relation is not
created at this point in time.

Now we just rely on the signal that is emitted when the anonymous user is
created and let CATMAID create the user profile like for the other users. This,
of course, only works if CATMAID's migrations are run *before* guardians. But
since guardian comes after CATMAID in the INSTALLED_APPLICATIONS list, this is
taken care of.

Thanks to Eric Perlman for bringing this to my attention.