Make user evaluation multi user review aware
To make the user evaluation admin tool support multi user reviews, this commit
makes it use always the newest review on a node to build epochs. Currently, all
reviews are taken into account, but this will be changed to make sure now "own"
reviews are included for the user being evaluated.

This change is part of the discussion in issue #580.