Remove reviewer_id and review_time fields from location based tables
With the new review system, these columns are now obsolete in the
treenode and the connector table. This commit removes this columns,
so make sure you have a backup!

The migration removes the fields only from the location table, because
the connector and the treenode table inherit it (as specified in the
initial migration). They will therefore also not have both fields
anymore after applying the migration

This is part of issue #580.