Statistics widget: add time unit selector
The time unit specifies an interval which is represented as an individual column
in the contribution record table. All values within this interval are
aggregated. All needed data is only retrieved once, so that the unit can be
changed without a reload. "Day" (default), "Week", "Month" and "Year" are the
available options. For simplicity, a month is expecte to be 30 days long. This
feature was suggested in issue #625.