Only request full message list if there are new messages
This commit changes the message polling to ask the back-end only for the date of
the most recent message. If there is a newer one than the one known on the
client side, the messages are actually retrieved. This will reduce the overhead,
when there are some unread messages. Now these are not retrieved over and over
again. The problem was stated in issue #643.