Neuron search widget: reset "sync to" select box on new search
We haven't really discussed whether this should be reset, but the current
behavior is not consistent. On a new search, the widget has still the
synchronization target selected that it had before. Selecting/deselecting the
new result also adds or removes them from the sync target. However, skeletons
that are already in the sync target are not removed and stay there. The
"sync to" functionality should make sure there are only those skeletons in the
target that are in the source list.

To make this more consistent, this commit resets the sync to select box on a
new search back to "None".