Annotations: keep annotator when joining and splitting skeletons
Up to now, the annotator of a kept annotations on joining and splitting
skeletons was lost. Instead of the original annotator, the current user
was used.  This is not the intended behavior, so this commit makes sure
the original annotator is kept on join and split operations.

Thanks to Ingrid Andrade for reporting this bug. This commit fixes #653.