Statistics widget: fix minor bug in pie legend generation
The legend of the pie chart will display all names in columns. This,
however, only worked for the first two columns. If there were more than
two columns required, only the first two were shown. This commit fixes
this problem.

This also fixes point 8 of the list of things to refurbish in the this
widget, discussed in issue #625. There, a pie chart entry for all
"Others" below the threshold was requested. It was actually there, but
the label wasn't displayed. Hence it wasn't recognizable.