Annotations: don't allow duplicates from different user on same entity
There is currently no need for allowing the same annotation to be linked
multiple times to the same entity by different users. Doing this complicates
things and also causes a problem when multiple "locked" annotations are added
(in this case, the entity is locked twice---who wins?). This commit makes sure
annotations can only be added once, regardless of the user. It also extends the
returned object to include the IDs of all entities that have actually be newly
annotated with a specific annotation.

This fixes #656, in which it was reported than one can add own "locked"
annotations to locked entities, thereby gaining permissions on them.