Navigator: fix display of usage time, annotator and usage count
Up to now all annotation lists displayed in the neuron navigator showed
very confusion information for the last usage time, the annotator and
the usage count for each annotation. These three columns always referred
to all instance of a particular annotation, it didn't respect user and
neuron filters. By not doing this, the output was usually misleading: If
one, for instance, added the 'locked' annotation to a neuron, the neuron
navigator would show for all occurrences of the locked annotation the
data for this last addition of one's 'locked' annotation, i.e. oneself
as annotator, its point of time as last used time and the overall usage
count. This is wrong and confusing.

This commit makes these three columns (last used, annotator, usage
count) respect user ID and neuron ID filters. It is clearly a bug that
this has not been done before. This change fixes #657.

Thanks to Javier Valdes Aleman for spotting this problem.